Rename 'src' to 'core' + group 'packages' + 'core' into a new 'src'
