+++ package.json
@@ -73,13 +73,13 @@
   ],
   "scripts": {
     "build": "rm -rf dist && tsc",
-    "postbuild": "cp src/*.html dist/src && cp packages/*.smash dist/packages",
+    "postbuild": "cp src/core/*.html dist/core && cp src/packages/*.smash dist/packages",
     "test:unit": "NODE_OPTIONS=\"${NODE_OPTIONS} --loader ts-node/esm\" nyc --reporter=lcov mocha tests/core/*.js tests/packages/*.js",
-    "test:browser:chrome": "ts-node --esm ./src/cli.js tests/packages/*.smash --groups=chrome",
+    "test:browser:chrome": "ts-node --esm ./src/core/cli.js tests/packages/*.smash --groups=chrome",
     "test": "npm run test:unit && npm run test:browser:chrome",
     "prepublishOnly": "eslint . && npm run test:unit && npm run build"
   },
   "bin": {
-    "smashtest": "./dist/src/cli.js"
+    "smashtest": "./dist/core/cli.js"
   }
 }
