+++ src/core/tree.js
