+++ tests/core/runner.js
@@ -1,9 +1,9 @@
 /* globals describe, it */
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import Comparer from '../../packages/js/comparer.js';
-import Runner from '../../src/runner.js';
-import Tree from '../../src/tree.js';
+import Comparer from '../../src/packages/js/comparer.js';
+import Runner from '../../src/core/runner.js';
+import Tree from '../../src/core/tree.js';
 
 const expect = chai.expect;
