+++ tests/packages/elementfinder.smash
@@ -3,8 +3,8 @@ open browser
     open test page + [
 
         inits {
-            await i('ElementFinder', '../../packages/js/elementfinder.js');
-            await i('Comparer', '../../packages/js/comparer.js');
+            await i('ElementFinder', '../../src/packages/js/elementfinder.js');
+            await i('Comparer', '../../src/packages/js/comparer.js');
             chai.use(i('chai-as-promised'));
 
             g('setPageBody', async (htmlStr) => {
